Separate input/output fps to avoid loosing buffered input, fixes #2
It is still not clear to me why the original design was problematic.

It doesn't seem like fflush(3) was the culprit, as removing it makes
no difference, we still throw away the entire input buffer after each
call to write(2). Looking at the fdopen(3) manpage, there are no hints
about issues with a read+write fp on a socket.

Accept reality, use a separate Rx and Tx fp, and move on to more
important things :)